refactor: tx updater (#1355)
